requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
contourpy==1.0.7
crashtest==0.4.1
cycler==0.11.0
Cython==0.29.33
dimod==0.12.4
discord==2.2.2
discord.py==2.2.2
diskcache==5.4.0
distlib==0.3.6
dnspython==2.4.2
dulwich==0.21.3
dwave-cloud-client==0.10.4
dwave-neal==0.6.0
dwave-networkx==0.8.13
dwave-preprocessing==0.5.4
dwave-samplers==1.0.0
dwave-system==1.18.0
eventlet==0.33.3
fasteners==0.18
filelock==3.10.0
Flask==2.2.3
fonttools==4.39.2
frozenlist==1.3.3
greenlet==2.0.2
homebase==1.0.1
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
kiwisolver==1.4.4
lockfile==0.12.2
MarkupSafe==2.1.2
matplotlib==3.7.1
minorminer==0.2.11
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
networkx==2.8.8
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pexpect==4.8.0
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==2.6.2
plucky==0.4.3
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
ptyprocess==0.7.0
py-sudoku==1.0.3
pycairo==1.23.0
pycparser==2.21
pydantic==1.10.7
PyGObject==3.42.2
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2022.7.1
rapidfuzz==2.13.7
rectangle-packer==2.0.1
requests==2.28.2
requests-toolbelt==0.10.1
scipy==1.10.1
seaborn==0.12.2
shellingham==1.5.0.post1
six==1.16.0
tomlkit==0.11.6
trove-classifiers==2023.3.9
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
webencodings==0.5.1
Werkzeug==2.2.3
xattr==0.10.1
yarl==1.8.2
zipp==3.15.0
